requirements.txt
asyncio
aiogram>=3.7.0
sqlalchemy>=2.0.30
aiosqlite
python-dotenv